fix error

+++ src/wallet/hdwallet.h
@@ -846,7 +846,7 @@ CHDWallet *GetVircleWallet(CKeyStore *win);
 const CHDWallet *GetVircleWallet(const CKeyStore *win);
 
 //add 2020-3-10
-void print_vec(const std::vector<int>& vec)
+void print_vec(const std::vector<uint8_t>& vec)
 {
     for (auto x: vec) {
          std::cout << ' ' << x;
